GH-284: [feat] CR refactor and add badges
